# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: iSHARE Onboarding API
  version: "1.0"
  description: Backend API for the iSHARE onboarding portal.
  contact:
    name: Protium.digital
    email: support@protium.digital
servers:
  - url: /
    description: Local server root
security:
  - bearerAuth: []
paths:
...
  /registry:
    get:
      tags: [registry]
      summary: List authorisation registries
      security: []
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, additionalProperties: true }}}}
        "500": { description: Internal Server Error, content: { application/json: { schema: { type: object, additionalProperties: {} }}}}
...
  /settings:
    get:
      tags: [settings]
      summary: Get settings
      security: []
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/models.Settings' }}}}
        "500": { description: Internal Server Error, content: { application/json: { schema: { type: object, additionalProperties: {} }}}}
    post:
      tags: [settings]
      summary: Update settings
      description: Requires onboarding-admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/models.Settings' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, additionalProperties: {} }}}}
        "400": { description: Bad Request, content: { application/json: { schema: { type: object, additionalProperties: {} }}}}
        "500": { description: Internal Server Error, content: { application/json: { schema: { type: object, additionalProperties: {} }}}}
components:
  schemas:
...
    models.Settings:
      type: object
      properties:
        agreements:
          type: array
          items:
            type: string
        dataspaceId:
          type: string
        description:
          type: string
        id:
          type: integer
        registrarId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT